refactor: simplify imports with index in Presentation
