+++ src/presentation/controllers/authentication/index.ts
@@ -1,0 +1,2 @@
+export * from './sign-in-controller'
+export * from './sign-in-controller'
