+++ src/presentation/controllers/authentication/sign-up-controller.ts
@@ -1,10 +1,7 @@
-import { type Controller } from '@/presentation/protocols/controller'
-import { type Validation } from '@/presentation/protocols/validation'
-import { type HttpResponse } from '@/presentation/protocols/http'
-import { badRequest, forbidden, ok, serverError } from '@/presentation/helpers/http-helper'
-import { EmailInUseError } from '@/presentation/errors/email-in-use-error'
-import { type AddAccount } from '@/domain/usecases/add-account'
-import { type Authentication } from '@/domain/usecases/authentication'
+import { type Controller, type Validation, type HttpResponse } from '@/presentation/protocols'
+import { badRequest, forbidden, ok, serverError } from '@/presentation/helpers'
+import { EmailInUseError } from '@/presentation/errors'
+import { type AddAccount, type Authentication } from '@/domain/usecases/authentication'
 
 export class SignUpController implements Controller {
   constructor(
