+++ src/presentation/errors/index.ts
@@ -1,0 +1,5 @@
+export * from './email-in-use-error'
+export * from './invalid-param-error'
+export * from './missing-param-error'
+export * from './server-error'
+export * from './unauthorized-error'
