+++ src/presentation/protocols/controller.ts
@@ -1,4 +1,4 @@
-import { type HttpResponse } from '@/presentation/protocols/http'
+import { type HttpResponse } from '@/presentation/protocols'
 
 export interface Controller<T = any> {
   handle: (request: T) => Promise<HttpResponse>
