+++ src/presentation/protocols/index.ts
@@ -1,0 +1,3 @@
+export * from './controller'
+export * from './http'
+export * from './validation'
